[S12P21A701-1156] fix: Add missing subscription_period field

+++ backend/src/main/java/bank/donghang/core/accountproduct/domain/AccountProduct.java
@@ -42,7 +42,7 @@ public class AccountProduct {
 	private AccountProductType accountProductType;
 
 	@Column(name = "subscription_period")
-	private Long subscriptionPeriod;
+	private Integer subscriptionPeriod;
 
 	@Column(name = "rate_description")
 	private String rateDescription;
